requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
ConfigSpace==0.6.1
bokeh==2.4.3
dask==2023.7.1
dask-jobqueue==0.8.2
distributed==2023.7.1
emcee==3.1.4
filelock==3.12.2
fsspec==2023.6.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.1
locket==1.0.0
MarkupSafe==2.1.3
more-itertools==10.0.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
partd==1.4.0
Pillow==10.0.0
psutil==5.9.5
pynisher==1.0.8
pyparsing==3.1.1
pyrfr==0.9.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smac==2.0.1
sortedcontainers==2.4.0
sympy==1.12
tblib==2.0.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
toolz==0.12.0
--extra-index-url https://download.pytorch.org/whl/cu118 torch==2.0.1+cu118
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
zict==3.0.0
zipp==3.16.2
